requirements.txt
anyio==3.6.1
asgiref==3.5.2
backports.cached-property==1.0.1
black==22.3.0
click==8.1.3
Django==4.0.5
graphql-core==3.2.1
h11==0.13.0
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
Pygments==2.12.0
python-dateutil==2.8.2
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.20.3
strawberry-graphql==0.114.1
tomli==2.0.1
typing_extensions==4.2.0
uvicorn==0.17.6
